requirements.txt
absl-py==0.8.0
bert4keras==0.7.2
joblib==0.14.1
numpy==1.17.0
scikit-learn==0.21.3
tensorflow>=2.0